% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mod.R
\name{add_mod_ui}
\alias{add_mod_ui}
\title{Fonction pour ajouter des menuItems dans le "sidebar", incluant des menuSubItem}
\usage{
add_mod_ui(my_dashboard, menu_title, my_tabName, ui = NULL, subitems = NULL)
}
\arguments{
\item{my_dashboard}{Fichier contenant les menu_items et tab_items sélectionnés}

\item{menu_title}{Titre qu'on veut voir apparaitre dans le "sidebar"}

\item{my_tabName}{Mot servant à faire le lien entre la selection dans le sidebar et le tabItem (une sorte de ID)}

\item{ui}{ui à ajouter}

\item{subitems}{subitem, introduit avec: list(add_mod_ui_sub(...),add_mod_ui_sub(...),...)}
}
\value{
Un menuItem aini que les les différents tabItem inclus dans celui-ci
}
\description{
Fonction pour ajouter des menuItems dans le "sidebar", incluant des menuSubItem
}
\examples{
if(interactive()){
add_mod_ui("my_dashboard","menu_title","my_tabName")
}
}
