% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mod.R
\name{add_mod_ui_sub}
\alias{add_mod_ui_sub}
\title{fct pour ajouter des modules à l'application}
\usage{
add_mod_ui_sub(condition, sub_title, my_tabName, ui)
}
\arguments{
\item{condition}{(boolean) Condition si cet item est ajouté ou pas au dashboard}

\item{sub_title}{Titre du menuSubItem dans le "sidebar"}

\item{my_tabName}{Mot servant à faire le lien entre la selection dans le sidebar et le tabItem (une sorte de ID)}

\item{ui}{section UI du module qu'on veut ajouter à l'application}
}
\value{
Ajoute un menu_sub_item et tab_item
}
\description{
fct pour ajouter des modules à l'application
}
\examples{
if(interactive()){
add_mod_ui_sub(condition, sub_title, my_tabName, ui)
}
}
