% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies_fcts.R
\name{cookie_actvity_timestamp}
\alias{cookie_actvity_timestamp}
\title{update cookie et cookie validator lors d'un input}
\usage{
cookie_actvity_timestamp(just_logged_out, input, session)
}
\arguments{
\item{just_logged_out}{reactiveVal servant à éviter le cookie auto-connect}

\item{input}{Variable input de la session shiny}

\item{session}{Variable de la session shiny}
}
\value{
Ne retourne rien, mais met à jour le cookie et cookie validator
}
\description{
update cookie et cookie validator lors d'un input
}
\examples{
if(interactive()){
cookie_actvity_timestamp(input, session)
}
}
