% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies_fcts.R
\name{cookie_auto_login}
\alias{cookie_auto_login}
\title{Reconnection automatique si fingerprint == cookie validator non expiré}
\usage{
cookie_auto_login(input, session)
}
\arguments{
\item{input}{variable input de la session shiny}

\item{session}{variable session de la session shiny}
}
\value{
Rien, mais modifie user_auth(), ce qui fait basculer dans l'application principale plutôt que la page d'authentification
}
\description{
Reconnection automatique si fingerprint == cookie validator non expiré
}
\examples{
if(interactive()){
cookie_auto_login(input,session)
}
}
