% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies_fcts.R
\name{cookie_remove_user}
\alias{cookie_remove_user}
\title{Fonction pour enregistrer un cookie}
\usage{
cookie_remove_user(session)
}
\arguments{
\item{session}{Paramètre de session shiny}
}
\value{
Ne retourne rien, mais efface le cookie
}
\description{
Principalement utiliser avec le bouton logout et inactivity
}
\examples{
if(interactive()){
cookie_remove_user(session)
}
}
