% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies_fcts.R
\name{cookie_set_user}
\alias{cookie_set_user}
\title{Fonction pour enregistrer un cookie}
\usage{
cookie_set_user(input, session)
}
\arguments{
\item{input}{input de la session shiny}

\item{session}{Paramètre de session shiny}
}
\value{
Ne retourne rien, mais enregistre un cookie et met à jour le "cookie validator" (fichier sur S3)
}
\description{
Fonction pour enregistrer un cookie
}
\examples{
if(interactive()){
cookie_set_user(input,session)
}
}
