% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies_fcts.R
\name{cookie_validator_delete}
\alias{cookie_validator_delete}
\title{Delete cookie validator}
\usage{
cookie_validator_delete(users, session)
}
\arguments{
\item{users}{(chr) vector avec les noms des users pour qui on veut effacer les cookie validators}

\item{session}{variable de la session shiny}
}
\value{
Rien mais efface les cookies validators sur S3
}
\description{
Delete cookie validator
}
\examples{
if(interactive()){
cookie_validator_delete(users,session)
}
}
