% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fingerprint.R
\name{fetch_client_ip}
\alias{fetch_client_ip}
\title{Fonction pour récupérer l'IP et le user agent
get user info}
\usage{
fetch_client_ip(session, ns = NULL, input_name = "client_ip_data")
}
\arguments{
\item{session}{variable session de shiny}

\item{ns}{Namespace, utile quand utilisé dans un module}

\item{input_name}{client_ip_data par défaut}
}
\value{
Rien mais ajoute un input
}
\description{
Fonction pour récupérer l'IP et le user agent
get user info
}
\examples{
if(interactive()){
fetch_client_ip(session)
}
}
