% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fingerprint.R
\name{get_fingerprint}
\alias{get_fingerprint}
\title{get code from ip and user agent}
\usage{
get_fingerprint(input)
}
\arguments{
\item{input}{input principal de la session shiny pour aller cherche information enregistré par nginx}
}
\value{
Un vecteur de type caractère représentant le fingerprint
}
\description{
get code from ip and user agent
}
\examples{
if(interactive()){
get_fingerprint(input)
}
}
