% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR_mod_config.R
\name{mod_config_ui1}
\alias{mod_config_ui1}
\title{Menu de configuration disponible à tous}
\usage{
mod_config_ui1(id)
}
\arguments{
\item{id}{ID du module}
}
\value{
UI de configuration pour l'utilisateur
}
\description{
Menu de configuration disponible à tous
}
\examples{
if(interactive()){
mod_config_ui1(id)
}
}
