% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR_mod_config.R
\name{mod_config_ui3}
\alias{mod_config_ui3}
\title{Menu de configuration disponible aux super_admin et dev}
\usage{
mod_config_ui3(id)
}
\arguments{
\item{id}{ID du module}
}
\value{
UI pour configuration de super_admin
}
\description{
Menu de configuration disponible aux super_admin et dev
}
\examples{
if(interactive()){
mod_config_ui3(id)
}
}
