% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR_mod_demo.R
\name{mod_demo2_ui}
\alias{mod_demo2_ui}
\title{module demo de protegR2, servant à présenter la structure + détail du role}
\usage{
mod_demo2_ui(id, session)
}
\arguments{
\item{id}{ID du module}

\item{session}{variable shiny avec info nécessaire au module}
}
\value{
UI demo
}
\description{
Particularité de ce module, en fait il ne s'agit que d'un fonction, car il n'y avait pas
de besoin au niveau du serveur
}
\examples{
if(interactive()){
mod_demo2_ui("demo2",session)
}
}
