% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_mod_demo.R
\name{mod_demo_airplane_ui}
\alias{mod_demo_airplane_ui}
\title{Module démo pour faire tourner un avion section ui}
\usage{
mod_demo_airplane_ui(id)
}
\arguments{
\item{id}{ID du module}
}
\value{
UI pour le module démo
}
\description{
Une particularité de ce module est qu'il utilise du java script.
Ce qui fait que c'est le navigateur de l'utilisateur qui fait tourner l'image
Ce qui évite de multiplier les calculs si jamais il y a beaucoup d'utilisateur
}
\examples{
if(interactive()){
mod_demo_airplane_ui(id)
}
}
