% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_mod_demo.R
\name{mod_demo_subitem1_server}
\alias{mod_demo_subitem1_server}
\title{Dummy module pour tester les subitems (section ui)}
\usage{
mod_demo_subitem1_server(id)
}
\arguments{
\item{id}{ID du module}
}
\value{
Calcul nécessaire au UI
}
\description{
Dummy module pour tester les subitems (section ui)
}
\examples{
if(interactive()){
mod_demo_subitem1_server(id)
}
}
