% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR_mod_demo.R
\name{mod_demo_subitem2}
\alias{mod_demo_subitem2}
\title{Dummy module pour tester les subitems}
\usage{
mod_demo_subitem2(id)
}
\arguments{
\item{id}{du module (ne serait pas nécessaire par il n'y a pas de section serveur au module)}
}
\value{
Un UI très simple
}
\description{
Dummy module pour tester les subitems
}
\examples{
if(interactive()){
mod_demo_subitem2(id)
}
}
