% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_mod_demo.R
\name{mod_fillable_server}
\alias{mod_fillable_server}
\title{Module démo pour illustrer page_fillable — section serveur}
\usage{
mod_fillable_server(id)
}
\arguments{
\item{id}{ID du module}
}
\value{
Calculs nécessaires au UI
}
\description{
Module démo pour illustrer page_fillable — section serveur
}
\examples{
if (interactive()) {
  mod_fillable_server("fillable_demo")
}
}
