% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_mod_demo.R
\name{mod_user_config_server}
\alias{mod_user_config_server}
\title{Module de test get_user_config / set_user_config — section serveur}
\usage{
mod_user_config_server(id)
}
\arguments{
\item{id}{ID du module}
}
\value{
Calculs nécessaires au UI
}
\description{
Module de test get_user_config / set_user_config — section serveur
}
