% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_login_logout.R
\name{perform_login}
\alias{perform_login}
\title{Fonction pour se connecter}
\usage{
perform_login(valid_user, token_value, input, session)
}
\arguments{
\item{valid_user}{Liste de liste avec les infos de l'utilisateur qui se connecte}

\item{token_value}{Token de la session à débuter}

\item{input}{input de la session shiny}

\item{session}{variable de la session shiny}
}
\value{
Ne retourne rien, mais enregistre un token et modifie l'utilisateur actif
}
\description{
Fonction pour se connecter
}
\examples{
if(interactive()){
perform_login(valid_user,token_value,session)
}
}
