% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_login_logout.R
\name{perform_logout}
\alias{perform_logout}
\title{Action lors du logout}
\usage{
perform_logout(session)
}
\arguments{
\item{session}{Variable de la session shiny}
}
\value{
Rien, mais modifie des paramètre
}
\description{
Efface le token et le cookie de la session et ensuite déconnecte l'utilisateur
}
\examples{
if(interactive()){
perform_logout(session)
}
}
