% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_mod_config_fcts.R
\name{protegR2_fct_validate_password}
\alias{protegR2_fct_validate_password}
\title{Fonction pour vérifier si deux passwords sont identiques (et minimum 5 caractère)}
\usage{
protegR2_fct_validate_password(pwd1, pwd2, min_length = 5)
}
\arguments{
\item{pwd1}{(chr) Password 1}

\item{pwd2}{(chr) Password 1}

\item{min_length}{(num) Nombre de caractère minimum pour le password}
}
\value{
True si identique et respecte les conditions; sinon FALSE
}
\description{
Fonction pour vérifier si deux passwords sont identiques (et minimum 5 caractère)
}
\examples{
if(interactive()){
protegR2_fct_validate_password(pwd1, pwd2, min_length = 5)
}
}
