% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_badge_button.R
\name{protegr2_update_badge}
\alias{protegr2_update_badge}
\title{Mettre a jour le badge d'un protegr2_badge_button}
\usage{
protegr2_update_badge(session = shiny::getDefaultReactiveDomain(), id, count)
}
\arguments{
\item{session}{Objet session Shiny. Par defaut la session courante.}

\item{id}{L'`id` passe a [protegr2_badge_button()].}

\item{count}{Nouveau compteur. `0` ou `NULL` pour masquer le badge.}
}
\description{
Modifie dynamiquement le compteur affiche sur un [protegr2_badge_button()].
}
