% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR_init.R
\name{protegR_copy_files}
\alias{protegR_copy_files}
\title{Copier les fichiers d'exemple du package vers l'application en construction}
\usage{
protegR_copy_files(background = FALSE, app = FALSE, R_files = FALSE)
}
\arguments{
\item{background}{Copier le fichier `background.png` vers `inst/app/www/`.}

\item{app}{Copier les fichier app.R server.R et global.R}

\item{R_files}{Copier les fichiers R (sauf `app.R`) vers le dossier `R/`.}
}
\value{
Aucun retour. Les fichiers sont copiés dans le projet courant.
}
\description{
Cette fonction copie certains fichiers inclus dans le package protegR`
vers l'arborescence d'une application Shiny en construction. Chaque groupe est contrôlé par un paramètre booléen.
Un message s'affiche si un fichier existe déjà pour demander s’il doit être écrasé.
}
\examples{
if (interactive()) {
  protegR_copy_files(background = TRUE, app = TRUE, R_files = TRUE)
}
}
