% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_mod_config.R
\name{mod_config_server}
\alias{mod_config_server}
\title{Section serveur servant du différents UI de configuration}
\usage{
mod_config_server(id, sessions, input_main_app, main_session)
}
\arguments{
\item{id}{ID du module}

\item{sessions}{variable globale contenant les infos de toutes les sessions actives}

\item{input_main_app}{Variable INPUT de la session shiny (inclus tous les inputs)}

\item{main_session}{Variable de la session shiny (plusieurs données commune)}
}
\value{
Retourne les calculs nécessaire aux différents UI de configuration
}
\description{
Section serveur servant du différents UI de configuration
}
\examples{
if(interactive()) {
mod_config_server(id,input_main_app,main_session)
}
}
