% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_mod_config.R
\name{mod_config_ui2}
\alias{mod_config_ui2}
\title{Menu de configuration disponible à tous sauf utilisateur de base}
\usage{
mod_config_ui2(id)
}
\arguments{
\item{id}{ID du module}
}
\value{
UI pour la configuration admin
}
\description{
Permet de créer de nouveaux utilisateurs et leur donner des accès
Un admin ne peux pas donner plus de privilège que ce qu'il détient lui-même
}
\examples{
if(interactive()){
mod_config_ui2(id)
}
}
