% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR_mod_config.R
\name{mod_config_ui4}
\alias{mod_config_ui4}
\title{Menu de configuration disponible aux dev}
\usage{
mod_config_ui4(id)
}
\arguments{
\item{id}{ID du module}
}
\value{
UI pour configuration accessible au dev
}
\description{
Menu de configuration disponible aux dev
}
\examples{
if(interactive()){
mod_config_ui4(id)
}
}
