% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR_mod_demo.R
\name{mod_demo1_server}
\alias{mod_demo1_server}
\title{module demo de protegR2, servant à présenter la structure, section serveur}
\usage{
mod_demo1_server(id)
}
\arguments{
\item{id}{ID du module}
}
\value{
Calcul nécessaire au UI
}
\description{
module demo de protegR2, servant à présenter la structure, section serveur
}
\examples{
if(interactive()){
mod_demo1_server("demo1")
}
}
