% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_mod_demo.R
\name{mod_demo1_ui}
\alias{mod_demo1_ui}
\title{module demo de protegR2, servant à présenter la structure}
\usage{
mod_demo1_ui(id, tr)
}
\arguments{
\item{id}{ID du module}

\item{tr}{Fonction de traduction retournée par make_tr()}
}
\value{
UI demo
}
\description{
module demo de protegR2, servant à présenter la structure
}
\examples{
if(interactive()){
mod_demo1_ui("demo1")
}
}
