% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR_mod_demo.R
\name{mod_demo_airplane_server}
\alias{mod_demo_airplane_server}
\title{Module démo pour faire tourner un avion section serveur}
\usage{
mod_demo_airplane_server(id)
}
\arguments{
\item{id}{ID du module}
}
\value{
fournit les calculs pour le module ui
}
\description{
Module démo pour faire tourner un avion section serveur
}
\examples{
if(interactive()){
mod_demo_airplane_server(id)
}
}
