% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR_mod_demo.R
\name{mod_demo_subitem1_ui}
\alias{mod_demo_subitem1_ui}
\title{Dummy module pour tester les subitems (section ui)}
\usage{
mod_demo_subitem1_ui(id)
}
\arguments{
\item{id}{du module}
}
\value{
Un UI très simple
}
\description{
Dummy module pour tester les subitems (section ui)
}
\examples{
if(interactive()){
mod_demo_subitem1(id)
}
}
