% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_mod_config_fcts.R
\name{protegR2_fct_change_pwd}
\alias{protegR2_fct_change_pwd}
\title{Fonction pour modifier un password}
\usage{
protegR2_fct_change_pwd(username, new_hash, config_s3_location)
}
\arguments{
\item{username}{Username pour lequel on veut changer le password}

\item{new_hash}{Password hashé}

\item{config_s3_location}{Information pour l'enregistrement sur S3}
}
\value{
Rien, mais modifie les données sur S3
}
\description{
Fonction pour modifier un password
}
\examples{
if(interactive()){
protegR2_fct_change_pwd(username, new_hash,config_s3_location)
}
}
