% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_init.R
\name{protegR2_init}
\alias{protegR2_init}
\title{Copier les fichiers d'aide du package vers le dossier dev/ du projet courant}
\usage{
protegR2_init()
}
\value{
Rien. Les fichiers sont copiés.
}
\description{
Copie tous les fichiers de `inst/files_to_copy/dev` (dans le package) vers le dossier `dev/`
du projet courant (là où la fonction est appelée).
}
\examples{
if (interactive()) protegR2_init()
}
