% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR_init.R
\name{protegR2_init_config_global}
\alias{protegR2_init_config_global}
\title{créer fichier initial config_global}
\usage{
protegR2_init_config_global(name, value)
}
\arguments{
\item{name}{nom du paramètre à ajouter (nom d'une liste)}

\item{value}{Valeur à insérer dans cette nouvelle liste}
}
\value{
Rien, sert à enregistrer un fichier de config_global sur S3
}
\description{
créer fichier initial config_global
}
\examples{
if(interactive()){
protegR2_init_config_global(cookie_name = "test")
}
}
