% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR2_init.R
\name{protegR2_init_record_s3_users_auth_file}
\alias{protegR2_init_record_s3_users_auth_file}
\title{Initialise un fichier d'authentification utilisateur sur S3}
\usage{
protegR2_init_record_s3_users_auth_file()
}
\value{
Rien. La fonction enregistre un fichier `.rds` sur S3.
}
\description{
Cette fonction génère un fichier `users_auth.rds` contenant des utilisateurs par défaut
avec des mots de passe hashés, puis l'enregistre sur un bucket S3.
}
\examples{
if (interactive()) {
  config_location <- list(
    s3_bucket = "mon-bucket",
    s3_main_folder = "mon-dossier"
  )
  protegR2_init_record_s3_users_auth_file()
}
}
