% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protegR_init.R
\name{protegR2_init_record_s3_users_auth_file_default}
\alias{protegR2_init_record_s3_users_auth_file_default}
\title{Version par défaut de `protegR2_init_record_s3_users_auth_file`}
\usage{
protegR2_init_record_s3_users_auth_file_default()
}
\value{
Rien.
}
\description{
Wrapper de la fonction principale. Lit les fichiers `config_s3_access.rds` et `config_s3_location.rds`
depuis le dossier `inst/app/data/`, puis appelle `protegR2_init3_record_s_users_auth_file()`.
}
\examples{
if (interactive()) {
  protegR2_init_record_s3_users_auth_file_default()
}
}
